+++ common/src/main/java/net/spell_engine/Platform.java
@@ -1,8 +1,9 @@
 package net.spell_engine;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
+import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.network.packet.Packet;
+import net.minecraft.server.network.ServerPlayerEntity;
 
 public class Platform {
     public static final boolean Fabric;
@@ -26,6 +27,10 @@ public class Platform {
     public interface Util {
         boolean isModLoaded(String modid);
         void awakeSlotModCompat();
+
+        // These are likely necessary due to some mapping differences
+        void sendVanillaPacket_C2S(ClientPlayerEntity player, Packet<?> packet);
+        void sendVanillaPacket_S2C(ServerPlayerEntity player, Packet<?> packet);
     }
 
     @ExpectPlatform