+++ fabric/src/main/java/net/spell_engine/fabric/PlatformImpl.java
@@ -1,6 +1,9 @@
 package net.spell_engine.fabric;
 
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.network.packet.Packet;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.spell_engine.Platform;
 import net.spell_engine.fabric.compat.FabricCompatFeatures;
 
@@ -18,6 +21,16 @@ public class PlatformImpl {
         @Override
         public void awakeSlotModCompat() {
             FabricCompatFeatures.initSlotCompat();
+        }
+
+        @Override
+        public void sendVanillaPacket_C2S(ClientPlayerEntity player, Packet<?> packet) {
+            player.networkHandler.sendPacket(packet);
+        }
+
+        @Override
+        public void sendVanillaPacket_S2C(ServerPlayerEntity player, Packet<?> packet) {
+            player.networkHandler.sendPacket(packet);
         }
     }
     private static final Platform.Util UTIL = new FabricUtil();